requirements.txt
# Core Deep Learning & ONNX Dependencies
tensorflow==2.15.0
tf2onnx==1.16.1
onnx==1.17.0
onnxruntime==1.23.2
onnx-graphsurgeon==0.5.8

# Axelera / NPU Specific
ethos-u-vela==4.4.1

# Image & Video Processing
opencv-python==4.9.0.80
pillow==12.0.0
scikit-image==0.25.2
imageio==2.37.2

# Math & Data manipulation
numpy==1.26.4
scipy==1.15.3
pandas==2.3.3

# Utilities and standard ML tooling
protobuf==3.20.3
h5py==3.15.1
tensorboard==2.15.2
networkx==3.4.2
matplotlib==3.10.7

# (Note: This file reflects the key dependencies of the 'vela' conda environment used for EdgeFlowNet deployment)
